Address more comments
