+++ rust/services/call/server/tests/test_helpers/gas_meter.rs
@@ -1,0 +1,14 @@
+use call_server::gas_meter::V_ALLOCATE_GAS;
+use derive_more::Deref;
+use serde::Serialize;
+use server_utils::RpcServerMock;
+
+#[derive(Deref)]
+pub struct ServerMock(RpcServerMock);
+
+impl ServerMock {
+    pub async fn start(params: impl Serialize, result: impl Serialize) -> ServerMock {
+        let mock_server = RpcServerMock::start(V_ALLOCATE_GAS, true, params, result).await;
+        Self(mock_server)
+    }
+}
